CODE/requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
opencv-python
pillow
scikit-learn
tensorflow